Split classes to their own files
