+++ GTAVMod_Speedometer/MaxSpeedState.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GTAVMod_Speedometer
+{
+    enum MaxSpeedState
+    {
+        Off = 0,
+        Counting = 1,
+    }
+}
