+++ GTAVMod_Speedometer/SpeedoMode.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GTAVMod_Speedometer
+{
+    enum SpeedoMode
+    {
+        Off = 0,
+        Simple = 1,
+        Detailed = 2,
+    }
+}
