+++ GTAVMod_Speedometer/UpdateCheckState.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GTAVMod_Speedometer
+{
+    enum UpdateCheckState
+    {
+        Stopped = 0,
+        Checking = 1,
+        Checked = 2,
+    }
+}
